requirements.txt
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.8.0
certifi==2026.2.25
cffi==2.0.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.4.4
constantly==23.10.4
cryptography==46.0.5
daphne==4.0.0
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.57.0
django-cors-headers==4.3.0
django-filter==23.5
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
graphene==3.4.3
graphene-django==3.1.5
graphql-core==3.2.7
graphql-relay==3.2.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
jwcrypto==1.5.6
msgpack==1.1.2
oauthlib==3.3.1
packaging==26.0
promise==2.3
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
PyMySQL==1.1.0
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2025.2
redis==7.2.1
requests==2.32.5
requests-oauthlib==2.0.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.3
zope.interface==8.2
